Add order management MCP tools
