+++ x10/tools/mcp/__init__.py
@@ -1,0 +1,3 @@
+from x10.tools.mcp.mcp_server import mcp as mcp_server
+
+__all__ = ["mcp_server"]
